refactor: admin challenge controller separation
