+++ src/main/java/com/example/green/domain/challenge/controller/docs/AdminChallengeQueryControllerDocs.java
@@ -1,0 +1,82 @@
+package com.example.green.domain.challenge.controller.docs;
+
+import com.example.green.domain.challenge.controller.dto.admin.AdminChallengeDetailResponseDto;
+import com.example.green.domain.challenge.controller.dto.admin.AdminChallengeParticipantListResponseDto;
+import com.example.green.domain.challenge.controller.dto.admin.AdminPersonalChallengeListResponseDto;
+import com.example.green.domain.challenge.controller.dto.admin.AdminTeamChallengeGroupDetailResponseDto;
+import com.example.green.domain.challenge.controller.dto.admin.AdminTeamChallengeGroupListResponseDto;
+import com.example.green.domain.challenge.controller.dto.admin.AdminTeamChallengeListResponseDto;
+import com.example.green.global.api.ApiTemplate;
+import com.example.green.global.api.page.CursorTemplate;
+import com.example.green.global.docs.ApiErrorStandard;
+import com.example.green.global.error.dto.ExceptionResponse;
+
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.enums.ParameterIn;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.tags.Tag;
+
+@Tag(name = "어드민 챌린지 관리 API", description = "어드민 챌린지 생성/수정/이미지/전시여부 등 관리 API")
+public interface AdminChallengeQueryControllerDocs {
+
+	@Operation(summary = "개인 챌린지 목록 조회", description = "개인 챌린지 목록을 조회합니다. (10개씩 조회)")
+	@ApiErrorStandard
+	@ApiResponse(responseCode = "200", description = "개인 챌린지 목록 조회 성공", useReturnTypeSchema = true)
+	@ApiResponse(responseCode = "403", description = "관리자 권한이 필요합니다.",
+		content = @Content(schema = @Schema(implementation = ExceptionResponse.class)))
+	ApiTemplate<CursorTemplate<Long, AdminPersonalChallengeListResponseDto>> getPersonalChallenges(
+		@Parameter(description = "커서 (마지막 챌린지 ID) - 첫 번째 조회 시에는 아무것도 넣지 말고 조회하세요") Long cursor
+	);
+
+	@Operation(summary = "팀 챌린지 목록 조회", description = "팀 챌린지 목록을 조회합니다. (10개씩 조회)")
+	@ApiErrorStandard
+	@ApiResponse(responseCode = "200", description = "팀 챌린지 목록 조회 성공", useReturnTypeSchema = true)
+	@ApiResponse(responseCode = "403", description = "관리자 권한이 필요합니다.",
+		content = @Content(schema = @Schema(implementation = ExceptionResponse.class)))
+	ApiTemplate<CursorTemplate<Long, AdminTeamChallengeListResponseDto>> getTeamChallenges(
+		@Parameter(description = "커서 (마지막 챌린지 ID) - 첫 번째 조회 시에는 아무것도 넣지 말고 조회하세요") Long cursor
+	);
+
+	@Operation(summary = "챌린지 상세 조회", description = "챌린지 상세 정보를 조회합니다.")
+	@ApiErrorStandard
+	@ApiResponse(responseCode = "200", description = "챌린지 상세 조회 성공", useReturnTypeSchema = true)
+	@ApiResponse(responseCode = "403", description = "관리자 권한이 필요합니다.",
+		content = @Content(schema = @Schema(implementation = ExceptionResponse.class)))
+	ApiTemplate<AdminChallengeDetailResponseDto> getChallengeDetail(
+		@Parameter(name = "challengeId", description = "챌린지 ID",
+			in = ParameterIn.PATH, required = true, example = "1") Long challengeId
+	);
+
+	@Operation(summary = "챌린지 참여자 목록 조회", description = "챌린지의 참여자 목록을 조회합니다. (10개씩 조회)")
+	@ApiErrorStandard
+	@ApiResponse(responseCode = "200", description = "참여자 목록 조회 성공", useReturnTypeSchema = true)
+	@ApiResponse(responseCode = "403", description = "관리자 권한이 필요합니다.",
+		content = @Content(schema = @Schema(implementation = ExceptionResponse.class)))
+	ApiTemplate<CursorTemplate<Long, AdminChallengeParticipantListResponseDto>> getChallengeParticipants(
+		@Parameter(name = "challengeId", description = "챌린지 ID",
+			in = ParameterIn.PATH, required = true, example = "1") Long challengeId,
+		@Parameter(description = "커서 (마지막 참여자 ID) - 첫 번째 조회 시에는 아무것도 넣지 말고 조회하세요") Long cursor
+	);
+
+	@Operation(summary = "그룹 목록 조회", description = "챌린지 그룹 목록을 조회합니다. (10개씩 조회)")
+	@ApiErrorStandard
+	@ApiResponse(responseCode = "200", description = "그룹 목록 조회 성공", useReturnTypeSchema = true)
+	@ApiResponse(responseCode = "403", description = "관리자 권한이 필요합니다.",
+		content = @Content(schema = @Schema(implementation = ExceptionResponse.class)))
+	ApiTemplate<CursorTemplate<Long, AdminTeamChallengeGroupListResponseDto>> getGroups(
+		@Parameter(description = "커서 (마지막 그룹 ID) - 첫 번째 조회 시에는 아무것도 넣지 말고 조회하세요") Long cursor
+	);
+
+	@Operation(summary = "그룹 상세 조회", description = "챌린지 그룹 상세 정보를 조회합니다.")
+	@ApiErrorStandard
+	@ApiResponse(responseCode = "200", description = "그룹 상세 조회 성공", useReturnTypeSchema = true)
+	@ApiResponse(responseCode = "403", description = "관리자 권한이 필요합니다.",
+		content = @Content(schema = @Schema(implementation = ExceptionResponse.class)))
+	ApiTemplate<AdminTeamChallengeGroupDetailResponseDto> getGroupDetail(
+		@Parameter(name = "groupId", description = "그룹 ID",
+			in = ParameterIn.PATH, required = true, example = "1") Long groupId
+	);
+}
